reformat code

+++ src/runner/executable.ts
@@ -18,13 +18,23 @@ interface BinaryMetadata {
     code: string;
 }
 
-export async function pushBinary(name: string, language: Language, code: string, path: string): Promise<void> {
+export async function pushBinary(
+    name: string,
+    language: Language,
+    code: string,
+    path: string
+): Promise<void> {
     winston.verbose(`Pushing binary ${name}, creating tar archive...`);
-    const binary = await streamToBuffer(tar.create({
-        gzip: true,
-        cwd: path,
-        portable: true
-    }, ['.']));
+    const binary = await streamToBuffer(
+        tar.create(
+            {
+                gzip: true,
+                cwd: path,
+                portable: true
+            },
+            ['.']
+        )
+    );
     const data: BinaryMetadata = {
         language: language.name,
         code: code
@@ -34,16 +44,20 @@ export async function pushBinary(name: string, language: Language, code: string,
 }
 
 // Return value: [path, language, code]
-export async function fetchBinary(name: string): Promise<[string, Language, string]> {
+export async function fetchBinary(
+    name: string
+): Promise<[string, Language, string]> {
     winston.verbose(`Fetching binary ${name}...`);
     await fse.ensureDir(Cfg.binaryDirectory);
     const targetName = pathLib.join(Cfg.binaryDirectory, name);
     const lockFileName = pathLib.join(Cfg.binaryDirectory, `${name}-get.lock`);
 
-    const metadata = msgpack.decode(await getRedis(name + redisMetadataSuffix)) as BinaryMetadata;
+    const metadata = msgpack.decode(
+        await getRedis(name + redisMetadataSuffix)
+    ) as BinaryMetadata;
     const isCurrentlyWorking = await fse.pathExists(lockFileName);
     // The binary already exists, no need for locking
-    if (await fse.pathExists(targetName) && !isCurrentlyWorking) {
+    if ((await fse.pathExists(targetName)) && !isCurrentlyWorking) {
         winston.debug(`Binary ${name} exists, no need for fetching...`);
     } else {
         winston.debug(`Acquiring lock ${lockFileName}...`);
@@ -59,7 +73,9 @@ export async function fetchBinary(name: string): Promise<[string, Language, stri
                 winston.debug(`Doing work: fetching binary for ${name} ...`);
                 await fse.ensureDir(targetName);
                 const binary = await getRedis(name + redisBinarySuffix);
-                winston.debug(`Decompressing binary (size=${binary.length})...`);
+                winston.debug(
+                    `Decompressing binary (size=${binary.length})...`
+                );
                 await new Promise((res, rej) => {
                     const s = tar.extract({
                         cwd: targetName
@@ -72,8 +88,7 @@ export async function fetchBinary(name: string): Promise<[string, Language, stri
             }
             ok = true;
         } finally {
-            if (!ok)
-                await fse.rmdir(targetName);
+            if (!ok) await fse.rmdir(targetName);
             winston.debug('Unlocking...');
             await lockfile.unlockAsync(lockFileName);
         }
